Refactored LrCalculateDimensions as a function
