+++ src/components/health/LrHealth.vue
@@ -16,7 +16,6 @@
 -->
 <template>
   <div class="health-page-container">
-    <LrCalulateDimensions />
     <SgtTabs :tabsInfo="tabsInfo" @selected="onTabChange" />
     <component :is="activeTab" />
   </div>
@@ -26,8 +25,6 @@ import SgtTabs, { TabsInfo } from "../shared/SgtTabs.vue";
 import { Component, Vue, Prop, Watch } from "vue-property-decorator";
 // import LrHealthTabular from "./LrHealthTabular.vue";
 import LrHealthGraphical from "./LrHealthGraphical.vue";
-import LrCalulateDimensions from "../LrCalulateDimensions.vue";
-import Tab1 from "./Tab1.vue";
 import Tab2 from "./Tab2.vue";
 
 @Component({
@@ -35,9 +32,7 @@ import Tab2 from "./Tab2.vue";
   components: {
     SgtTabs,
     LrHealthGraphical,
-    LrCalulateDimensions,
     // LrHealthTabular,
-    Tab1,
     Tab2,
   },
 })
@@ -52,13 +47,11 @@ export default class LrHealth extends Vue {
       label: "Tabular",
     },
   ];
-  // private activeTab = "Tab1";
   private activeTab = "LrHealthGraphical";
 
   public onTabChange(value: number) {
     switch (value) {
       case 1:
-        // this.activeTab = "Tab1";
         this.activeTab = "LrHealthGraphical";
         break;
       case 2:
