--- src/components/health/Tab1.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>Tab1</div>
-</template>
-
-<script>
-export default {
-  name: "Tab1",
-};
-</script>
-
-<style></style>
